+++ talker/example/talker_example.dart
@@ -22,22 +22,18 @@ Future<void> main() async {
   try {
     throw Exception('Test service exception');
   } catch (e, st) {
-    Talker.instance.handle(e, 'Working with string error', st);
+    Talker.instance.handle(e, st, 'Working with string error');
   }
 
   try {
     throw Exception('Service can`t get test data');
   } on Exception catch (e, st) {
-    Talker.instance.handleException(e, 'Working with strings exception', st);
+    Talker.instance.handleException(e, st, 'Working with strings exception');
   }
 
   Talker.instance.log(
     'Server error',
     logLevel: LogLevel.critical,
-    additional: {
-      "status": 500,
-      "error": "Internal Server Error",
-    },
   );
 
   Talker.instance.fine('Log fine');