+++ talker/lib/src/observers/observers_manager.dart
@@ -1,23 +1,29 @@
 import 'package:talker/src/models/talker_data/talker_data.dart';
 import 'package:talker/src/observers/talker_observer.dart';
-import 'package:talker_error_handler/talker_error_handler.dart';
 
 /// Manager to wrap all observers
 class TalkerObserversManager {
   TalkerObserversManager(this.observers);
   final List<TalkerObserver> observers;
 
-  /// Called when [Talker] handle an error / exception
-  void onError(ErrorDetails container) {
+  /// Called when [Talker] handle an [TalkerError]
+  void onError(TalkerError talkerError) {
     for (final o in observers) {
-      o.onError?.call(container);
+      o.onError?.call(talkerError);
     }
   }
 
-  /// Called when [Talker] handle an log
-  void onLog(TalkerDataInterface data) {
+  /// Called when [Talker] handle an [TalkerException]
+  void onException(TalkerException talkerException) {
     for (final o in observers) {
-      o.onLog?.call(data);
+      o.onException?.call(talkerException);
+    }
+  }
+
+  /// Called when [Talker] handle an [TalkerDataInterface] log
+  void onLog(TalkerDataInterface log) {
+    for (final o in observers) {
+      o.onLog?.call(log);
     }
   }
 }