+++ talker/lib/src/utils/error_handler/error_handler.dart
@@ -1,0 +1,2 @@
+export 'talker_error_handler.dart';
+export 'talker_error_handler_interface.dart';