+++ talker/lib/talker.dart
@@ -1,6 +1,6 @@
 library talker;
 
-export 'package:talker_error_handler/talker_error_handler.dart';
+// export 'package:talker_error_handler/talker_error_handler.dart';
 export 'package:talker_logger/talker_logger.dart';
 
 export 'src/src.dart';